+++ gui/seleccion_robot.py
@@ -99,9 +99,11 @@ class PantallaSelecRobot(Pantalla):
         pass
 
     def eliminar_robot(self, indice):
-        self.controlador.eliminar_robot(indice)
-        self.tabla.tabla.delete(self.tabla.tabla.focus())
-        self.tabla.desactivar_botones()
+        if tk.messagebox.askyesno("Eliminar?",
+                                  "Eliminar robot y todos sus modelos?"):
+            self.controlador.eliminar_robot(indice)
+            self.tabla.tabla.delete(self.tabla.tabla.focus())
+            self.tabla.desactivar_botones()
 
     def actualizar(self):
         super().actualizar()
